refactor(nexus-nwju): move retry helpers to nexus.retry leaf module
- Create src/nexus/retry.py as a true leaf module (stdlib + httpx + structlog only)
- Move _is_retryable_chroma_error, _chroma_with_retry, _is_retryable_voyage_error,
  _voyage_with_retry from nexus.db.t3 to nexus.retry
- Update all callers: doc_indexer.py, indexer.py, scoring.py now import from nexus.retry
- Remove local import workaround in scoring.py rerank_results (top-level import now safe)
- Update time.sleep patches in test_chroma_retry.py and test_voyage_retry.py
  from nexus.db.t3.time → nexus.retry.time
- Extend test_no_circular_imports to verify nexus.retry has no nexus.db attribute

References: nexus-nwju
