+++ src/nexus/db/t3.py
@@ -5,10 +5,8 @@ from __future__ import annotations
 import hashlib
 import os
 import threading
-import time
-from collections.abc import Callable
 from datetime import UTC, datetime, timedelta
-from typing import Any, Literal
+from typing import Literal
 
 import chromadb
 import chromadb.errors
@@ -28,115 +26,12 @@ _log = structlog.get_logger(__name__)
 # nexus_rdr, nexus_knowledge.
 _STORE_TYPES: tuple[str, ...] = ("code", "docs", "rdr", "knowledge")
 
-
-# ── ChromaDB transient-error retry ───────────────────────────────────────────
-
-_RETRYABLE_FRAGMENTS: frozenset[str] = frozenset({
-    "502", "503", "504", "429",
-    "bad gateway", "service unavailable", "gateway time-out", "too many requests",
-})
-_RETRYABLE_HTTP_STATUSES: frozenset[int] = frozenset({429, 502, 503, 504})
-
-
-def _is_retryable_chroma_error(exc: BaseException) -> bool:
-    """Return True if *exc* represents a transient ChromaDB Cloud error worth retrying.
-
-    Check order:
-    1. Transport-level errors (ConnectError, ReadTimeout, RemoteProtocolError) — always retry.
-    2. Chained httpx.HTTPStatusError — authoritative integer status code check.
-    3. String fallback — plain Exception message body (gateway HTML or chroma JSON).
-    """
-    # 1. Transport-level errors — no HTTP response, but clearly transient.
-    if isinstance(exc, httpx.TransportError):
-        return True
-    # 2. ChromaDB wraps HTTPStatusError as Exception(resp.text); original is __context__.
-    ctx = exc.__context__
-    if isinstance(ctx, httpx.HTTPStatusError):
-        return ctx.response.status_code in _RETRYABLE_HTTP_STATUSES
-    # 3. Fallback: scan the message body for retryable status tokens.
-    msg = str(exc).lower()
-    return any(fragment in msg for fragment in _RETRYABLE_FRAGMENTS)
-
-
-def _chroma_with_retry(
-    fn: Callable[..., Any],
-    *args: Any,
-    max_attempts: int = 5,
-    **kwargs: Any,
-) -> Any:
-    """Call *fn* with exponential backoff on transient ChromaDB Cloud errors.
-
-    Retries up to *max_attempts* times (default 5).  Backoff starts at 2 s,
-    doubles each attempt, capped at 30 s.  Non-retryable errors raise immediately.
-    """
-    delay = 2.0
-    for attempt in range(1, max_attempts + 1):
-        try:
-            return fn(*args, **kwargs)
-        except Exception as exc:
-            if attempt == max_attempts or not _is_retryable_chroma_error(exc):
-                raise
-            _log.warning(
-                "chroma_transient_error_retry",
-                attempt=attempt,
-                delay=delay,
-                error=str(exc)[:120],
-            )
-            time.sleep(delay)
-            delay = min(delay * 2, 30.0)
-
-
-
-# ── Voyage AI transient-error retry ──────────────────────────────────────────
-
-try:
-    import voyageai.error as _voyageai_error
-    _VOYAGE_ERROR_TYPES: tuple[type, ...] | None = (
-        _voyageai_error.APIConnectionError,
-        _voyageai_error.TryAgain,
-    )
-except ImportError:  # pragma: no cover
-    _VOYAGE_ERROR_TYPES = None
-
-
-def _is_retryable_voyage_error(exc: BaseException) -> bool:
-    """Return True if *exc* is a transient Voyage AI error worth retrying.
-
-    Only APIConnectionError and TryAgain are retried here.  Timeout,
-    RateLimitError, and ServiceUnavailableError are handled by the built-in
-    ``max_retries`` on ``voyageai.Client`` (tenacity-based).  The two error
-    spaces are disjoint; do not add Voyage AI types to _is_retryable_chroma_error.
-    """
-    return bool(_VOYAGE_ERROR_TYPES and isinstance(exc, _VOYAGE_ERROR_TYPES))
-
-
-def _voyage_with_retry(
-    fn: Callable[..., Any],
-    *args: Any,
-    max_attempts: int = 3,
-    **kwargs: Any,
-) -> Any:
-    """Call *fn* with backoff on transient Voyage AI errors (APIConnectionError, TryAgain).
-
-    Retries up to *max_attempts* times (default 3).  Backoff starts at 1 s,
-    doubles each attempt, capped at 10 s.  Non-retryable errors raise immediately.
-    """
-    delay = 1.0
-    for attempt in range(1, max_attempts + 1):
-        try:
-            return fn(*args, **kwargs)
-        except Exception as exc:
-            if attempt == max_attempts or not _is_retryable_voyage_error(exc):
-                raise
-            _log.warning(
-                "voyage_transient_error_retry",
-                attempt=attempt,
-                delay=delay,
-                error=str(exc)[:120],
-            )
-            time.sleep(delay)
-            delay = min(delay * 2, 10.0)
-
+from nexus.retry import (
+    _chroma_with_retry,
+    _is_retryable_chroma_error,
+    _is_retryable_voyage_error,
+    _voyage_with_retry,
+)
 
 class T3Database:
     """T3 ChromaDB CloudClient permanent knowledge store.
